pyproject.toml
[tool.poetry]
name = "bl-damage-classifier"
version = "0.1.0"
description = ""
authors = ["Filip-231 <filip.szmid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
openai = "^1.27.0"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
pandas = "^2.2.2"
pytesseract = "^0.3.10"
opencv-python = "^4.10.0.82"
easyocr = "^1.7.1"
google-cloud-vision = "^3.7.2"
matplotlib = "^3.9.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
openpyxl = "^3.1.3"
loguru = "^0.7.2"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pytest = "^8.2.2"
httpx = "^0.27.0"
scikit-learn = "^1.5.1"
tensorboard = "^2.17.0"
black = "^24.4.2"
random-word = "^1.0.11"
tensorflow = "^2.17.0"
hyperopt = "^0.2.7"
motor = "^3.5.1"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
